1. 불필요한 파일 삭제 2. package 분리
