--- android/app/src/main/java/com/example/moHaeng/GoogleMapsFragment.kt
@@ -1,75 +1,0 @@
-package com.example.moHaeng
-
-import android.content.Context
-import androidx.fragment.app.Fragment
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-import com.google.android.gms.maps.CameraUpdateFactory
-import com.google.android.gms.maps.GoogleMap
-import com.google.android.gms.maps.MapView
-import com.google.android.gms.maps.OnMapReadyCallback
-import com.google.android.gms.maps.model.LatLng
-import com.google.android.gms.maps.model.MarkerOptions
-
-class GoogleMapsFragment : Fragment(), OnMapReadyCallback {
-
-    lateinit var mContext: Context
-
-    override fun onAttach(context: Context) {
-        super.onAttach(context)
-//
-//        if (context is BackActivity) {
-//            mContext = context
-//        }
-    }
-    private lateinit var mView: MapView
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        var rootView = inflater.inflate(R.layout.fragment_maps, container, false)
-        mView = rootView.findViewById(R.id.mapView)
-        mView.onCreate(savedInstanceState)
-        mView.getMapAsync(this)
-        return rootView
-    }
-
-    override fun onMapReady(googleMap: GoogleMap) {
-        val marker = LatLng(37.568291,126.997780)
-        googleMap.addMarker(MarkerOptions().position(marker).title("여기"))
-        googleMap.moveCamera(CameraUpdateFactory.newLatLng(marker))
-        googleMap.moveCamera(CameraUpdateFactory.zoomTo(15f))
-    }
-
-    override fun onStart() {
-        super.onStart()
-        mView.onStart()
-    }
-    override fun onStop() {
-        super.onStop()
-        mView.onStop()
-    }
-    override fun onResume() {
-        super.onResume()
-        mView.onResume()
-    }
-    override fun onPause() {
-        super.onPause()
-        mView.onPause()
-    }
-    override fun onLowMemory() {
-        super.onLowMemory()
-        mView.onLowMemory()
-    }
-    override fun onDestroy() {
-        mView.onDestroy()
-        super.onDestroy()
-    }
-
-}
