--- android/app/src/main/java/com/example/moHaeng/MyPageActivity.kt
@@ -1,11 +1,0 @@
-package com.example.moHaeng
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MyPageActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_my_page)
-    }
-}
