--- android/app/src/main/java/com/example/moHaeng/RankingItem.kt
@@ -1,5 +1,0 @@
-package com.example.moHaeng
-
-class RankingItem {
-
-}
