+++ android/app/src/main/java/com/example/moHaeng/productSearch/RankingItem.kt
@@ -1,0 +1,5 @@
+package com.example.moHaeng.productSearch
+
+class RankingItem {
+
+}
