+++ android/app/src/main/java/com/example/moHaeng/sort/SortItem.kt
@@ -1,4 +1,4 @@
-package com.example.moHaeng
+package com.example.moHaeng.sort
 
 class SortItem(s: String, b: Boolean) {
     val sortName = s
