+++ android/app/src/main/java/com/example/moHaeng/sort/SortedFragment.kt
@@ -1,4 +1,4 @@
-package com.example.moHaeng
+package com.example.moHaeng.sort
 
 import android.os.Bundle
 import android.view.LayoutInflater
