+++ android/app/src/main/res/layout/fragment_add_item.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddEventFragment">
+    tools:context=".addEvent.AddEventFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/add_event_toolbar"
