+++ android/app/src/main/res/layout/fragment_sorted.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:paddingEnd="16dp"
     android:paddingStart="16dp"
-    tools:context=".SortedFragment">
+    tools:context=".sort.SortedFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="wrap_content"
